refactor(usage): move analytics read queries to domain layer (AYC-334)
The LocalUsageRepository (a driven infrastructure adapter) imported four
query classes from its own application/use-cases layer, inverting the
hexagonal dependency direction flagged by the adapters-no-use-cases
dependency-cruiser rule.

These query objects are part of the UsageRepository read contract, so
move GetProviderUsageQuery, GetModelDistributionQuery, GetUserUsageQuery
and GetUsageStatsQuery into usage/domain/queries/ — a location the port,
adapter, use cases and presenters can all depend on without violating any
architecture rule. Update all imports and drop the four grandfathered
adapters-no-use-cases entries from the known-violations baseline.

Also extract the verbose Swagger @ApiQuery decorators on the user-usage
endpoint into a composed ApiUserUsageQueries() decorator so the handler
stays within the max-lines-per-function complexity budget once the file
is touched.
