+++ ayunis-core-backend/src/domain/usage/application/ports/usage.repository.ts
@@ -1,9 +1,9 @@
 import type { UUID } from 'crypto';
 import type { Usage } from '../../domain/usage.entity';
-import type { GetUserUsageQuery } from '../use-cases/get-user-usage/get-user-usage.query';
-import type { GetProviderUsageQuery } from '../use-cases/get-provider-usage/get-provider-usage.query';
-import type { GetModelDistributionQuery } from '../use-cases/get-model-distribution/get-model-distribution.query';
-import type { GetUsageStatsQuery } from '../use-cases/get-usage-stats/get-usage-stats.query';
+import type { GetUserUsageQuery } from '../../domain/queries/get-user-usage.query';
+import type { GetProviderUsageQuery } from '../../domain/queries/get-provider-usage.query';
+import type { GetModelDistributionQuery } from '../../domain/queries/get-model-distribution.query';
+import type { GetUsageStatsQuery } from '../../domain/queries/get-usage-stats.query';
 import type { Paginated } from 'src/common/pagination';
 import { UsageStats } from '../../domain/usage-stats.entity';
 import { ProviderUsage } from '../../domain/provider-usage.entity';
