+++ ayunis-core-backend/src/domain/usage/application/use-cases/get-model-distribution/get-model-distribution.use-case.spec.ts
@@ -1,7 +1,7 @@
 import type { TestingModule } from '@nestjs/testing';
 import { Test } from '@nestjs/testing';
 import { GetModelDistributionUseCase } from './get-model-distribution.use-case';
-import { GetModelDistributionQuery } from './get-model-distribution.query';
+import { GetModelDistributionQuery } from '../../../domain/queries/get-model-distribution.query';
 import { UsageRepository } from '../../ports/usage.repository';
 import { InvalidDateRangeError } from '../../usage.errors';
 import { ModelDistribution } from 'src/domain/usage/domain/model-distribution.entity';
