+++ ayunis-core-backend/src/domain/usage/application/use-cases/get-provider-usage/get-provider-usage.use-case.spec.ts
@@ -1,7 +1,7 @@
 import type { TestingModule } from '@nestjs/testing';
 import { Test } from '@nestjs/testing';
 import { GetProviderUsageUseCase } from './get-provider-usage.use-case';
-import { GetProviderUsageQuery } from './get-provider-usage.query';
+import { GetProviderUsageQuery } from '../../../domain/queries/get-provider-usage.query';
 import { UsageRepository } from '../../ports/usage.repository';
 import { InvalidDateRangeError } from '../../usage.errors';
 import { ProviderUsage } from 'src/domain/usage/domain/provider-usage.entity';
