+++ ayunis-core-backend/src/domain/usage/application/use-cases/get-usage-stats/get-usage-stats.use-case.ts
@@ -1,5 +1,5 @@
 import { Injectable, Logger } from '@nestjs/common';
-import { GetUsageStatsQuery } from './get-usage-stats.query';
+import { GetUsageStatsQuery } from '../../../domain/queries/get-usage-stats.query';
 import { UsageRepository } from '../../ports/usage.repository';
 import { UsageStats } from 'src/domain/usage/domain/usage-stats.entity';
 import { validateOptionalDateRange } from '../../usage.utils';
