+++ ayunis-core-backend/src/domain/usage/application/use-cases/get-user-usage/get-user-usage.use-case.ts
@@ -1,5 +1,5 @@
 import { Injectable, Logger } from '@nestjs/common';
-import { GetUserUsageQuery } from './get-user-usage.query';
+import { GetUserUsageQuery } from '../../../domain/queries/get-user-usage.query';
 import {
   UsageRepository,
   type UserUsageResult,
