+++ ayunis-core-backend/src/domain/usage/presenters/http/super-admin-usage.controller.ts
@@ -16,8 +16,8 @@ import { ConfigService } from '@nestjs/config';
 import { parseDate } from './utils/parse-date.util';
 import { UUID } from 'crypto';
 import { UsageUseCasesFacade } from './usage-use-cases.facade';
-import { GetModelDistributionQuery } from '../../application/use-cases/get-model-distribution/get-model-distribution.query';
-import { GetUsageStatsQuery } from '../../application/use-cases/get-usage-stats/get-usage-stats.query';
+import { GetModelDistributionQuery } from '../../domain/queries/get-model-distribution.query';
+import { GetUsageStatsQuery } from '../../domain/queries/get-usage-stats.query';
 import { UsageConstants } from '../../domain/value-objects/usage.constants';
 import { SystemRoles } from 'src/iam/authorization/application/decorators/system-roles.decorator';
 import { SystemRole } from 'src/iam/users/domain/value-objects/system-role.enum';
